update docs, FromStr for IdBuf
